Use git instead of 'builtin' for base vcpkg repository
Using builtin leads to misleading error messages, unless my user
installation of vcpkg happens to be at the exact correct tag.

As well as the misleading error messages, depending on my per-user
versions seems underirable, as it conflicts with the goals of manifest
mode.

This also switches to a full commit hash, as that is required for `git`
repositories; it seems like an oversight/accident that it isn't for
`builtin`. In general, tags can be moved, so aren't suitable when
supply-chain-attacks are plausible